replace parcel with vite
