--- react-example-init/.postcssrc
@@ -1,5 +1,0 @@
-{
-  "plugins": {
-    "tailwindcss": {}
-  }
-}
