+++ react-example-init/index.html
@@ -7,6 +7,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="./src/index.js"></script>
+    <script type="module" src="./src/index.jsx"></script>
   </body>
 </html>
