+++ react-example-init/package.json
@@ -1,19 +1,20 @@
 {
-  "name": "react-example-payment",
+  "name": "react-example-init",
   "version": "1.0.0",
   "description": "",
   "scripts": {
-    "start": "npx parcel index.html"
+    "start": "vite"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "parcel": "^2.13.2",
-    "postcss": "^8.4.49",
-    "process": "^0.11.10",
+    "@tailwindcss/vite": "^4.1.17",
+    "@vitejs/plugin-react": "^5.1.2",
+    "@vitejs/plugin-react-swc": "^4.2.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "tailwindcss": "^3.4.3"
+    "tailwindcss": "^4.1.17",
+    "vite": "^7.2.7"
   },
   "dependencies": {
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
