+++ react-example-init/src/authorize-split/index.jsx
