+++ react-example-init/src/index.jsx
