+++ react-example-init/styles/index.css
@@ -1,6 +1,4 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
+@import "tailwindcss";
 
 h1 {
   @apply text-4xl;
