+++ react-example-initTokenize/index.html
@@ -7,6 +7,6 @@
   <body>
     <h1>Secure Fields for PCI Proxy customers</h1>
     <div id="app"></div>
-    <script type="module" src="./src/index.js"></script>
+    <script type="module" src="./src/index.jsx"></script>
   </body>
 </html>
