+++ react-example-initTokenize/package.json
@@ -1,18 +1,20 @@
 {
-  "name": "react-example-pciproxy",
+  "name": "react-example-initTokenize",
   "version": "1.0.0",
   "description": "",
   "scripts": {
-    "start": "npx parcel index.html"
+    "start": "vite"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@tailwindcss/vite": "^4.1.17",
+    "@vitejs/plugin-react": "^5.1.2",
+    "@vitejs/plugin-react-swc": "^4.2.2",
     "classnames": "^2.5.1",
-    "parcel": "^2.12.0",
-    "process": "^0.11.10",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "vite": "^7.2.7"
   },
   "dependencies": {
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
