+++ react-example-initTokenize/src/SecureFields.jsx
