+++ react-example-initTokenize/src/index.jsx
